# docs/openapi.json
{
  "components": {
    "schemas": {
...
      "DisputeResponse": {
        "description": "A dispute's state.",
        "properties": {
          "id": {
            "title": "Id",
            "type": "string"
          },
          "order_id": {
            "title": "Order Id",
            "type": "string"
          },
          "reason": {
            "title": "Reason",
            "type": "string"
          },
          "resolution_note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resolution Note"
          },
          "status": {
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "id",
          "order_id",
          "status",
          "reason"
        ],
        "title": "DisputeResponse",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "RefreshRequest": {
        "additionalProperties": false,
        "description": "Exchange a refresh token for a new token pair.",
        "properties": {
          "refresh_token": {
            "description": "The current refresh token.",
            "title": "Refresh Token",
            "type": "string"
          }
        },
        "required": [
          "refresh_token"
        ],
        "title": "RefreshRequest",
        "type": "object"
      },
...
      "RegisterRequest": {
        "additionalProperties": false,
        "description": "Create an account authorised by a registration token.",
        "properties": {
          "city": {
            "anyOf": [
              {
                "maxLength": 100,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Legacy city name for couriers; prefer city_id from GET /api/cities.",
            "title": "City"
          },
          "city_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "City Id"
          },
          "dob": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dob"
          },
          "email": {
            "anyOf": [
              {
                "maxLength": 255,
                "pattern": "^[^@\\s]+@[^@\\s]+\\.[^@\\s]+$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Email"
          },
          "full_name": {
            "anyOf": [
              {
                "maxLength": 120,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Full Name"
          },
          "gender": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/UserGender"
              },
              {
                "type": "null"
              }
            ]
          },
          "national_id": {
            "anyOf": [
              {
                "maxLength": 64,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Courier identity (national id or passport required).",
            "title": "National Id"
          },
          "passport_id": {
            "anyOf": [
              {
                "maxLength": 64,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Passport Id"
          },
          "registration_token": {
            "description": "From verify-otp when is_new_user is true.",
            "title": "Registration Token",
            "type": "string"
          },
          "role": {
            "description": "The account role to create.",
            "enum": [
              "CUSTOMER",
              "COURIER"
            ],
            "title": "Role",
            "type": "string"
          }
        },
        "required": [
          "registration_token",
          "role"
        ],
        "title": "RegisterRequest",
        "type": "object"
      },
      "RejectWithdrawalRequest": {
        "additionalProperties": false,
        "description": "An admin rejection reason for a withdrawal.",
        "properties": {
          "reason": {
            "maxLength": 255,
            "minLength": 1,
            "title": "Reason",
            "type": "string"
          }
        },
        "required": [
          "reason"
        ],
        "title": "RejectWithdrawalRequest",
        "type": "object"
      },
      "ResolveDisputeRequest": {
        "additionalProperties": false,
        "description": "Admin resolution of a dispute.",
        "properties": {
          "courier_amount": {
            "anyOf": [
              {
                "maxLength": 20,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Courier's share for a SPLIT resolution (decimal string).",
            "title": "Courier Amount"
          },
          "note": {
            "anyOf": [
              {
                "maxLength": 1000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          },
          "outcome": {
            "description": "RESOLVED_CUSTOMER | RESOLVED_COURIER | RESOLVED_SPLIT.",
            "title": "Outcome",
            "type": "string"
          }
        },
        "required": [
          "outcome"
        ],
        "title": "ResolveDisputeRequest",
        "type": "object"
      },
...
      "SendOtpRequest": {
        "additionalProperties": false,
        "description": "Request an OTP for a phone number.",
        "properties": {
          "phone": {
            "description": "Saudi mobile; accepts 0501234567, 501234567, or +966501234567.",
            "examples": [
              "0501234567"
            ],
            "pattern": "^\\+9665\\d{8}$",
            "title": "Phone",
            "type": "string"
          }
        },
        "required": [
          "phone"
        ],
        "title": "SendOtpRequest",
        "type": "object"
      },
      "SendOtpResponse": {
        "description": "OTP request accepted.",
        "properties": {
          "expires_in": {
            "description": "Seconds until the OTP expires.",
            "examples": [
              60
            ],
            "title": "Expires In",
            "type": "integer"
          },
          "otp_dev": {
            "anyOf": [
              {
                "maximum": 99999.0,
                "minimum": 10000.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "description": "Five-digit OTP returned only in development; omitted elsewhere.",
            "title": "Otp Dev"
          }
        },
        "required": [
          "expires_in"
        ],
        "title": "SendOtpResponse",
        "type": "object"
      },
...
      "TokenResponse": {
        "description": "An issued access + refresh token pair.",
        "properties": {
          "access_token": {
            "description": "30-minute access JWT.",
            "title": "Access Token",
            "type": "string"
          },
          "refresh_token": {
            "description": "30-day rotating refresh token.",
            "title": "Refresh Token",
            "type": "string"
          },
          "role": {
            "description": "The authenticated user's role.",
            "title": "Role",
            "type": "string"
          }
        },
        "required": [
          "access_token",
          "refresh_token",
          "role"
        ],
        "title": "TokenResponse",
        "type": "object"
      },
...
      "UserGender": {
        "description": "Optional self-reported gender.",
        "enum": [
          "MALE",
          "FEMALE",
          "OTHER",
          "PREFER_NOT_TO_SAY"
        ],
        "title": "UserGender",
        "type": "string"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
      "VerifyOtpRequest": {
        "additionalProperties": false,
        "description": "Verify a submitted OTP.",
        "properties": {
          "otp": {
            "description": "Five digits in development; six digits in test and production.",
            "examples": [
              "849201"
            ],
            "pattern": "^\\d{5,6}$",
            "title": "Otp",
            "type": "string"
          },
          "phone": {
            "description": "Saudi mobile; accepts 0501234567, 501234567, or +966501234567.",
            "pattern": "^\\+9665\\d{8}$",
            "title": "Phone",
            "type": "string"
          }
        },
        "required": [
          "phone",
          "otp"
        ],
        "title": "VerifyOtpRequest",
        "type": "object"
      },
      "VerifyOtpResponse": {
        "description": "The result of verifying an OTP: tokens for an existing user, else a handoff.",
        "properties": {
          "access_token": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "30-minute access JWT, if existing user.",
            "title": "Access Token"
          },
          "is_new_user": {
            "description": "True when no account exists for the phone.",
            "title": "Is New User",
            "type": "boolean"
          },
          "refresh_token": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "30-day refresh token, if existing user.",
            "title": "Refresh Token"
          },
          "registration_token": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Short-lived token to call /api/auth/register, if a new user.",
            "title": "Registration Token"
          },
          "role": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "The user's role, when tokens are issued.",
            "title": "Role"
          }
        },
        "required": [
          "is_new_user"
        ],
        "title": "VerifyOtpResponse",
        "type": "object"
      },
...
      "WithdrawalResponse": {
        "description": "A masked courier withdrawal record.",
        "properties": {
          "amount": {
            "title": "Amount",
            "type": "string"
          },
          "iban_last4": {
            "title": "Iban Last4",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "rejection_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rejection Reason"
          },
          "status": {
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "id",
          "amount",
          "iban_last4",
          "status"
        ],
        "title": "WithdrawalResponse",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "Giftly API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/admin/disputes/{dispute_id}/resolve": {
      "post": {
        "description": "Resolve a dispute, moving escrow accordingly (ADMIN only).",
        "operationId": "resolve_dispute_api_admin_disputes__dispute_id__resolve_post",
        "parameters": [
          {
            "in": "path",
            "name": "dispute_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Dispute Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResolveDisputeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DisputeResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Resolve Dispute",
        "tags": [
          "admin-api"
        ]
      }
    },
    "/api/admin/withdrawals/{withdrawal_id}/approve": {
      "post": {
        "description": "Approve a requested withdrawal while retaining its funds hold.",
        "operationId": "approve_withdrawal_api_admin_withdrawals__withdrawal_id__approve_post",
        "parameters": [
          {
            "in": "path",
            "name": "withdrawal_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Withdrawal Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WithdrawalResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Approve Withdrawal",
        "tags": [
          "admin-api"
        ]
      }
    },
    "/api/admin/withdrawals/{withdrawal_id}/paid": {
      "post": {
        "description": "Settle an approved external payout through the append-only ledger.",
        "operationId": "mark_withdrawal_paid_api_admin_withdrawals__withdrawal_id__paid_post",
        "parameters": [
          {
            "in": "path",
            "name": "withdrawal_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Withdrawal Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WithdrawalResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Mark Withdrawal Paid",
        "tags": [
          "admin-api"
        ]
      }
    },
    "/api/admin/withdrawals/{withdrawal_id}/reject": {
      "post": {
        "description": "Reject a requested or approved withdrawal and release its funds.",
        "operationId": "reject_withdrawal_api_admin_withdrawals__withdrawal_id__reject_post",
        "parameters": [
          {
            "in": "path",
            "name": "withdrawal_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Withdrawal Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RejectWithdrawalRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WithdrawalResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Reject Withdrawal",
        "tags": [
          "admin-api"
        ]
      }
    },
...
    "/api/auth/refresh": {
      "post": {
        "description": "Rotate a refresh token, returning a new pair (reuse revokes the family).",
        "operationId": "refresh_api_auth_refresh_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Refresh",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/auth/register": {
      "post": {
        "description": "Create the account authorised by a registration token and return tokens.",
        "operationId": "register_api_auth_register_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Register",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/auth/send-otp": {
      "post": {
        "description": "Send a login OTP. The response is identical whether or not the phone exists.",
        "operationId": "send_otp_api_auth_send_otp_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SendOtpRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SendOtpResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Send Otp",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/auth/verify-otp": {
      "post": {
        "description": "Verify an OTP, returning tokens for an existing user or a registration token.",
        "operationId": "verify_otp_api_auth_verify_otp_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VerifyOtpRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VerifyOtpResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Verify Otp",
        "tags": [
          "auth"
        ]
      }
    },
...
  }
}